move component to table folder
