+++ src/client/components/Table/TableMobileView.tsx
@@ -13,10 +13,10 @@ import { ChevronForwardIcon } from '@amsterdam/design-system-react-icons';
 import classNames from 'classnames';
 
 import styles from './TableMobileView.module.scss';
+import type { TableV2Props } from './TableV2.types.ts';
+import { useDisplayPropsEntries } from './useDisplayPropEntries.hook.ts';
 import type { LinkProps } from '../../../universal/types/App.types.ts';
 import { MaLink, MaRouterLink } from '../MaLink/MaLink.tsx';
-import type { TableV2Props } from '../Table/TableV2.types.ts';
-import { useDisplayPropsEntries } from '../Table/useDisplayPropEntries.hook.ts';
 
 export function LinkOrFragment({
   children,
