+++ src/client/components/Thema/ThemaPaginaTable.tsx
@@ -10,9 +10,9 @@ import { MAX_TABLE_ROWS_ON_THEMA_PAGINA } from '../../config/app.ts';
 import { useSmallScreen } from '../../hooks/media.hook.ts';
 import { LinkToListPage } from '../LinkToListPage/LinkToListPage.tsx';
 import { PageContentCell } from '../Page/Page.tsx';
+import { TableMobileView } from '../Table/TableMobileView.tsx';
 import type { DisplayProps } from '../Table/TableV2.tsx';
 import { TableV2 } from '../Table/TableV2.tsx';
-import { TableMobileView } from '../TableMobileView/TableMobileView.tsx';
 
 const DISPLAY_PROPS_DEFAULT: DisplayProps<{ title: string }> = {
   title: 'Titel',
